+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Moves docs site packages and scripts to the docs directory (#57)
 - Removes Travis, adds GitHub Action (#58)
 - Components can accept either a controller or target element (#65, #70)
+- Deprecates MenuBar component; use Menu instead (#75)
+- Removes antipatterns from Menu component (#75)
 
 **Added**
 
@@ -27,6 +29,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `instance.on()` and `instance.off()` methods for subscribing to events (#67)
 - Menu, MenuBar, and MenuButton will log a warning if used for website navigation (#68)
 - Components track attributes they add and will only overwrite existing attribute values where required (#73)
+- Adds an autoClose option to Disclosure and Menu (#75, #76)
 
 **Fixed**
 
@@ -35,13 +38,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - MenuBar wasn't tracking the current Popup (6c2fe90)
 - MenuBar keydown could trigger an unexpected Popup `stateChange` event (e86d06e, 8001226)
 - Dialog no longer re-queries for interactive child elements on every TAB keydown (a964674)
+- Corrects an issue where shift-tab from the Popup controller would focus the target's first child (#76)
 
 **Removed**
 
 - Dialog no longer requires a close button, and will not create one (#51)
-- Components no longer accept `onInit`, `onStateChange`, nor `onDestroy` callbacks; additionally, MenuBar no longer acepts `onPopupInit` (#54, #55)
+- Components no longer accept `onInit`, `onStateChange`, nor `onDestroy` callbacks (#54, #55)
 - MenuBar no longer tracks Popup `expanded` state separately (e86d06e)
 - Components no longer manage the `hidden` attribute (#74)
+- MenuButton component (#75)
+- Remove `collapse` and `itemMatches` Menu options (#75)
 
 ## 0.3.1